Fix: fire `onControlFinished` for touch drags (#35)
Previously, `onControlFinished` was only invoked at the end of a mouse
operation (i.e. `mouseup` and `mouseleave`). This worked ok for _taps_,
at least on mobile Firefox, because the browser synthesizes a
`mouseup` event after the tap.

However, when dragging, browsers don't seem to synthesize this event.
Therefore, we should explicitly call `onControlFinished` on `touchend`
and `touchcancel`.